Address review: strip eachyear at set op instead of in displace
Rather than stripping the "eachyear" per-year suffix defensively inside the
displacement last-specified logic, strip it at the source. SetOperation did
not normalize away "eachyear" the way Cap/Change/Floor operations do, so a
baseline declared with `set ... each year` left "unitseachyear" in its
last-specified value. That called into the displacement conversion, which
rejects per-year units as a destination.

Strip "eachyear" in SetOperation (mirroring the other operations) so
last-specified values stay in canonical base units and internal engine
operations need no special handling. Simplify the displacement last-specified
conversion to always convert to the prior value's units (per review), keeping
the "sales" metastream in sync when a component stream is displaced.

Fold the new regression test into CapLiveTests and remove the standalone
CapLiveTest class.
