+++ engine/src/main/java/org/kigalisim/engine/support/StreamUpdateShortcuts.java
@@ -239,14 +239,13 @@ public class StreamUpdateShortcuts {
     if (EngineSupportUtils.getIsSalesStream(stream, false)) {
       UseKey destKey = new SimpleUseKey(destinationScope.getApplication(),
                                         destinationScope.getSubstance());
-      // Preserve the units of the last-specified value this is overwriting (e.g. keep equipment
-      // units instead of switching to kg), so unit-based carry-over / percentage-change semantics
-      // are not silently changed by displacement.
+      // Preserve the units of the last-specified value this is overwriting, so unit-based
+      // carry-over / percentage-change semantics are not silently changed by displacement.
       EngineNumber prior = engine.getStreamKeeper().getLastSpecifiedValue(destKey, stream);
       EngineNumber valueToRecord = value;
-      if (prior != null && prior.hasEquipmentUnits() && !value.hasEquipmentUnits()) {
+      if (prior != null) {
         UnitConverter unitConverter = EngineSupportUtils.createUnitConverterWithTotal(engine, stream);
-        valueToRecord = unitConverter.convert(value, "units");
+        valueToRecord = unitConverter.convert(value, prior.getUnits());
       }
       engine.getStreamKeeper().setLastSpecifiedValue(destKey, stream, valueToRecord);
     }
