+++ engine/src/test/java/org/kigalisim/validate/CapLiveTests.java
@@ -947,4 +947,59 @@ public class CapLiveTests {
         "Virgin should be higher when recycling is limited (was "
         + limitRecycleVirginYear2 + " vs BAU " + bauVirginYear2 + ")");
   }
+
+  /**
+   * Test a cap-and-displace scenario with recharge where the displaced substance starts with
+   * zero sales. SubA is capped (displacing to SubB) during years 3 to 4.
+   *
+   * <p>Displacement writes the displaced volume onto SubB's sales last-specified value (in the
+   * same units as the value it overwrites), so SubB's sales carry over after the cap window ends
+   * rather than collapsing back to zero in year 5.</p>
+   */
+  @Test
+  public void testSubConsumptionCarriesOverAfterCapWindow() throws IOException {
+    String qtaPath = "../examples/cap_displace_sales_drop_year5.qta";
+    ParsedProgram program = KigaliSimFacade.parseAndInterpret(qtaPath);
+    assertNotNull(program, "Program should not be null");
+
+    String scenarioName = "With Permit";
+    Stream<EngineResult> results = KigaliSimFacade.runScenario(program, scenarioName, progress -> {});
+    List<EngineResult> resultsList = results.collect(Collectors.toList());
+
+    // QTA names: application "Test", substances "SubA" and "SubB".
+    EngineResult subA3 = LiveTestsUtil.getResult(resultsList.stream(), 3, "Test", "SubA");
+    EngineResult subB3 = LiveTestsUtil.getResult(resultsList.stream(), 3, "Test", "SubB");
+    EngineResult subB4 = LiveTestsUtil.getResult(resultsList.stream(), 4, "Test", "SubB");
+
+    assertNotNull(subA3, "Should have result for Test/SubA in year 3");
+    assertNotNull(subB3, "Should have result for Test/SubB in year 3");
+    assertNotNull(subB4, "Should have result for Test/SubB in year 4");
+
+    // During the cap window (years 3-4) SubB receives displaced sales from SubA,
+    // so its consumption is substantial.
+    double subB3cons = subB3.getDomestic().getValue().doubleValue();
+    double subB4cons = subB4.getDomestic().getValue().doubleValue();
+    assertTrue(subB3cons > 0, "SubB consumption should be > 0 in year 3 (displacement active)");
+    assertTrue(subB4cons > 0, "SubB consumption should be > 0 in year 4 (displacement active)");
+
+    // Once the cap window ends (year 5), SubB must carry over the displaced sales
+    // rather than collapsing to only its own recharge: it continues selling new
+    // equipment and its consumption stays around (or above) the displaced level.
+    EngineResult subB5 = LiveTestsUtil.getResult(resultsList.stream(), 5, "Test", "SubB");
+    EngineResult subB6 = LiveTestsUtil.getResult(resultsList.stream(), 6, "Test", "SubB");
+    assertNotNull(subB5, "Should have result for Test/SubB in year 5");
+    assertNotNull(subB6, "Should have result for Test/SubB in year 6");
+
+    double subB5cons = subB5.getDomestic().getValue().doubleValue();
+    double subB6cons = subB6.getDomestic().getValue().doubleValue();
+    double subB5popNew = subB5.getPopulationNew().getValue().doubleValue();
+    assertTrue(subB5popNew > 0,
+        "SubB should keep adding new equipment in year 5 (displaced sales carry over)");
+    assertTrue(subB5cons >= subB4cons,
+        "SubB consumption should not drop in year 5 after the cap window ends; displaced "
+            + "sales carry over (year 4: " + subB4cons + ", year 5: " + subB5cons + ")");
+    assertTrue(subB6cons >= subB4cons,
+        "SubB consumption should remain at the carried-over displaced level in year 6 "
+            + "(year 6: " + subB6cons + ")");
+  }
 }
